Move FPS limiter into separate file
